moved export/import, added database reset function
